+++ src/pilot.h
@@ -206,7 +206,7 @@ typedef enum EscortType_e {
 typedef struct Escort_s {
    char *ship;          /**< Type of the ship escort is flying. */
    EscortType_t type;   /**< Type of escort. */
-   unsigned int id;     /**< ID of in-game pilot. */
+   unsigned long id; /**< ID of in-game pilot. */
    /* TODO: something better than this */
    int persist;         /**< True if escort should respawn on takeoff/landing */
 } Escort_t;
@@ -217,7 +217,7 @@ typedef struct Escort_s {
  */
 typedef struct Pilot_ {
 
-   unsigned int id;  /**< pilot's id, used for many functions */
+   unsigned long id; /**< pilot's id, used for many functions */
    char* name;       /**< pilot's name (if unique) */
 
    /* Fleet/faction management. */
@@ -322,16 +322,16 @@ typedef struct Pilot_ {
    PilotHook *hooks; /**< Array (array.h): Pilot hooks. */
 
    /* Escort stuff. */
-   unsigned int parent; /**< Pilot's parent. */
+   unsigned long parent; /**< Pilot's parent. */
    Escort_t *escorts; /**< Array (array.h): Pilot's escorts. */
-   unsigned int dockpilot; /**< Pilot's dock pilot (the pilot it originates from). This is
-                          separate from parent because it needs to be set in sync with
-                          dockslot (below). Used to unset dockslot when the dock pilot
-                          is destroyed. */
+   unsigned long dockpilot; /**< Pilot's dock pilot (the pilot it
+         originates from). This is separate from parent because it needs
+         to be set in sync with dockslot (below). Used to unset dockslot
+         when the dock pilot is destroyed. */
    int dockslot; /**< Outfit slot pilot originates from, index of dockpilot's outfits. */
 
    /* Targeting. */
-   unsigned int target; /**< AI pilot target. */
+   unsigned long target; /**< AI pilot target. */
    int nav_planet;   /**< Planet land target. */
    int nav_hyperspace; /**< Hyperspace target. */
    int nav_anchor; /**< Asteroid anchor target. */
@@ -381,17 +381,20 @@ typedef struct Pilot_ {
  * getting pilot stuff
  */
 Pilot*const* pilot_getAll (void);
-Pilot* pilot_get( const unsigned int id );
-unsigned int pilot_getNextID( const unsigned int id, int mode );
-unsigned int pilot_getPrevID( const unsigned int id, int mode );
-unsigned int pilot_getNearestEnemy( const Pilot* p );
-unsigned int pilot_getNearestEnemy_size( const Pilot* p, double target_mass_LB, double target_mass_UB );
-unsigned int pilot_getNearestEnemy_heuristic(const Pilot* p, double mass_factor, double health_factor, double damage_factor, double range_factor);
-unsigned int pilot_getNearestHostile (void); /* only for the player */
-unsigned int pilot_getNearestPilot( const Pilot* p );
-unsigned int pilot_getBoss( const Pilot* p );
-double pilot_getNearestPos( const Pilot *p, unsigned int *tp, double x, double y, int disabled );
-double pilot_getNearestAng( const Pilot *p, unsigned int *tp, double ang, int disabled );
+Pilot* pilot_get(const unsigned long id);
+unsigned long pilot_getNextID(const unsigned long id, int mode);
+unsigned long pilot_getPrevID(const unsigned long id, int mode);
+unsigned long pilot_getNearestEnemy(const Pilot* p);
+unsigned long pilot_getNearestEnemy_size(const Pilot* p, double target_mass_LB,
+      double target_mass_UB);
+unsigned long pilot_getNearestEnemy_heuristic(const Pilot* p,
+      double mass_factor, double health_factor, double damage_factor,
+      double range_factor);
+unsigned long pilot_getNearestHostile(void); /* only for the player */
+unsigned long pilot_getNearestPilot(const Pilot* p);
+unsigned long pilot_getBoss(const Pilot* p);
+double pilot_getNearestPos(const Pilot *p, unsigned long *tp, double x, double y, int disabled );
+double pilot_getNearestAng( const Pilot *p, unsigned long *tp, double ang, int disabled );
 int pilot_getJumps( const Pilot* p );
 const glColour* pilot_getColour( const Pilot* p );
 int pilot_validTarget( const Pilot* p, const Pilot* target );
@@ -404,10 +407,10 @@ double pilot_relhp( const Pilot* cur_pilot, const Pilot* p );
 /*
  * Combat.
  */
-void pilot_setTarget( Pilot* p, unsigned int id );
-double pilot_hit( Pilot* p, const Solid* w, const unsigned int shooter,
-      const Damage *dmg, int reset );
-void pilot_updateDisable( Pilot* p, const unsigned int shooter );
+void pilot_setTarget(Pilot* p, unsigned long id);
+double pilot_hit(Pilot* p, const Solid* w, const unsigned long shooter,
+      const Damage *dmg, int reset);
+void pilot_updateDisable(Pilot* p, const unsigned long shooter);
 void pilot_explode( double x, double y, double radius, const Damage *dmg, const Pilot *parent );
 double pilot_face( Pilot* p, const double dir );
 int pilot_brake( Pilot* p );
@@ -442,9 +445,10 @@ PilotOutfitSlot* pilot_getDockSlot( Pilot* p );
 /*
  * creation
  */
-unsigned int pilot_create( const Ship* ship, const char* name, int faction, const char *ai,
-      const double dir, const Vector2d* pos, const Vector2d* vel,
-      const PilotFlags flags, unsigned int dockpilot, int dockslot );
+unsigned long pilot_create(const Ship* ship, const char* name, int faction,
+      const char *ai, const double dir, const Vector2d* pos,
+      const Vector2d* vel, const PilotFlags flags, unsigned long dockpilot,
+      int dockslot);
 Pilot* pilot_createEmpty( const Ship* ship, const char* name,
       int faction, const char *ai, PilotFlags flags );
 Pilot* pilot_replacePlayer( Pilot* after );
@@ -485,7 +489,8 @@ void pilot_renderOverlay( Pilot* p, const double dt );
 /*
  * communication
  */
-void pilot_message( Pilot *p, unsigned int target, const char *msg, int ignore_int );
+void pilot_message(Pilot *p, unsigned long target, const char *msg,
+      int ignore_int);
 void pilot_broadcast( Pilot *p, const char *msg, int ignore_int );
 void pilot_distress(Pilot *p, Pilot *attacker, const char *msg);
